// src/static/enums.ts
export enum PrintColors {
  HEADER = "\033[95m",
  OKBLUE = "\033[94m",
  OKCYAN = "\033[96m",
  OKGREEN = "\033[92m",
  WARNING = "\033[93m",
  FAIL = "\033[91m",
  ENDC = "\033[0m",
  BOLD = "\033[1m",
  UNDERLINE = "\033[4m"
}

export enum Game {
  GENSHIN = "Genshin Impact",
  STARRAIL = "Honkai: Star Rail",
  ZZZ = "Zenless Zone Zero"
}

export enum GenshinCity {
  MONDSTADT = "Mondstadt",
  LIYUE = "Liyue",
  INAZUMA = "Inazuma",
  SUMERU = "Sumeru",
  FONTAINE = "Fontaine",
  NATLAN = "Natlan"
}